move reporters and callbacks to flattened logging module
