+++ graphrag/api/index_api.py
@@ -12,11 +12,9 @@ from graphrag.config import CacheType, GraphRagConfig
 from graphrag.index.cache.noop_pipeline_cache import NoopPipelineCache
 from graphrag.index.create_pipeline_config import create_pipeline_config
 from graphrag.index.emit.types import TableEmitterType
-from graphrag.index.progress import (
-    ProgressReporter,
-)
 from graphrag.index.run import run_pipeline_with_config
 from graphrag.index.typing import PipelineRunResult
+from graphrag.logging import ProgressReporter
 
 
 async def build_index(
