+++ graphrag/index/input/load_input.py
@@ -12,11 +12,11 @@ import pandas as pd
 
 from graphrag.config import InputConfig, InputType
 from graphrag.index.config import PipelineInputConfig
-from graphrag.index.progress import NullProgressReporter, ProgressReporter
 from graphrag.index.storage import (
     BlobPipelineStorage,
     FilePipelineStorage,
 )
+from graphrag.logging import NullProgressReporter, ProgressReporter
 
 from .csv import input_type as csv
 from .csv import load as load_csv
