--- graphrag/index/progress/__init__.py
@@ -1,18 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""Progress-reporting components."""
-
-from .types import (
-    NullProgressReporter,
-    PrintProgressReporter,
-    ProgressReporter,
-    ReporterType,
-)
-
-__all__ = [
-    "NullProgressReporter",
-    "PrintProgressReporter",
-    "ProgressReporter",
-    "ReporterType",
-]
