+++ graphrag/index/run/run.py
@@ -21,10 +21,6 @@ from graphrag.index.config import (
 )
 from graphrag.index.emit import TableEmitterType, create_table_emitters
 from graphrag.index.load_pipeline_config import load_pipeline_config
-from graphrag.index.progress import NullProgressReporter, ProgressReporter
-from graphrag.index.reporting import (
-    ConsoleWorkflowCallbacks,
-)
 from graphrag.index.run.cache import _create_cache
 from graphrag.index.run.postprocess import (
     _create_postprocess_steps,
@@ -52,6 +48,11 @@ from graphrag.index.workflows import (
     VerbDefinitions,
     WorkflowDefinitions,
     load_workflows,
+)
+from graphrag.logging import (
+    ConsoleWorkflowCallbacks,
+    NullProgressReporter,
+    ProgressReporter,
 )
 from graphrag.utils.storage import _create_storage
