+++ graphrag/index/run/workflow.py
@@ -17,13 +17,10 @@ from datashaper import (
 
 from graphrag.index.context import PipelineRunContext
 from graphrag.index.emit.table_emitter import TableEmitter
-from graphrag.index.progress.types import ProgressReporter
-from graphrag.index.reporting.progress_workflow_callbacks import (
-    ProgressWorkflowCallbacks,
-)
 from graphrag.index.run.profiling import _write_workflow_stats
 from graphrag.index.storage.typing import PipelineStorage
 from graphrag.index.typing import PipelineRunResult
+from graphrag.logging import ProgressReporter, ProgressWorkflowCallbacks
 from graphrag.utils.storage import _load_table_from_storage
 
 log = logging.getLogger(__name__)
