+++ graphrag/logging/console_status.py
@@ -1,26 +1,11 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-"""Status Reporter for orchestration."""
+"""Console Reporter."""
 
-from abc import ABCMeta, abstractmethod
 from typing import Any
 
-
-class StatusReporter(metaclass=ABCMeta):
-    """Provides a way to report status updates from the pipeline."""
-
-    @abstractmethod
-    def error(self, message: str, details: dict[str, Any] | None = None):
-        """Report an error."""
-
-    @abstractmethod
-    def warning(self, message: str, details: dict[str, Any] | None = None):
-        """Report a warning."""
-
-    @abstractmethod
-    def log(self, message: str, details: dict[str, Any] | None = None):
-        """Report a log."""
+from .types import StatusReporter
 
 
 class ConsoleStatusReporter(StatusReporter):
