+++ graphrag/logging/console_workflow_callbacks.py
