+++ graphrag/logging/load_progress_reporter.py
@@ -3,10 +3,10 @@
 
 """Load a progress reporter."""
 
-from .rich import RichProgressReporter
+from .null_progress import NullProgressReporter
+from .print_progress import PrintProgressReporter
+from .rich_progress import RichProgressReporter
 from .types import (
-    NullProgressReporter,
-    PrintProgressReporter,
     ProgressReporter,
     ReporterType,
 )
