+++ graphrag/query/llm/oai/chat_openai.py
@@ -15,13 +15,13 @@ from tenacity import (
     wait_exponential_jitter,
 )
 
+from graphrag.logging import StatusReporter
 from graphrag.query.llm.base import BaseLLM, BaseLLMCallback
 from graphrag.query.llm.oai.base import OpenAILLMImpl
 from graphrag.query.llm.oai.typing import (
     OPENAI_RETRY_ERROR_TYPES,
     OpenaiApiType,
 )
-from graphrag.query.progress import StatusReporter
 
 _MODEL_REQUIRED_MSG = "model is required"
