+++ graphrag/query/llm/oai/embedding.py
@@ -18,6 +18,7 @@ from tenacity import (
     wait_exponential_jitter,
 )
 
+from graphrag.logging import StatusReporter
 from graphrag.query.llm.base import BaseTextEmbedding
 from graphrag.query.llm.oai.base import OpenAILLMImpl
 from graphrag.query.llm.oai.typing import (
@@ -25,7 +26,6 @@ from graphrag.query.llm.oai.typing import (
     OpenaiApiType,
 )
 from graphrag.query.llm.text_utils import chunk_text
-from graphrag.query.progress import StatusReporter
 
 
 class OpenAIEmbedding(BaseTextEmbedding, OpenAILLMImpl):
